move changes to align-deps
