+++ packages/align-deps/scripts/update-profile.mjs
@@ -20,7 +20,7 @@ import { isMetaPackage } from "../lib/capabilities.js";
  *   name: string;
  *   version: string;
  *   latest: string;
- *   homepage: string;
+ *   homepage?: string;
  *   dependencies?: Record<string, string>;
  *   peerDependencies?: Record<string, string>;
  * }} PackageInfo
@@ -62,7 +62,13 @@ function getPackageVersion(packageName, dependencies) {
   if (!packageVersion) {
     throw new Error(`Failed to get '${packageName}' version`);
   }
-  return semverCoerce(packageVersion).version;
+
+  const coercedVersion = semverCoerce(packageVersion);
+  if (!coercedVersion) {
+    throw new Error(`Failed to coerce version: ${packageVersion}`);
+  }
+
+  return coercedVersion.version;
 }
 
 /**
@@ -103,6 +109,10 @@ function generateFromTemplate({
   metroVersion,
 }) {
   const nextVersionCoerced = semverCoerce(targetVersion);
+  if (!nextVersionCoerced) {
+    throw new Error(`Failed to coerce version: ${targetVersion}`);
+  }
+
   const currentVersion = `${nextVersionCoerced.major}.${
     nextVersionCoerced.minor - 1
   }`;
@@ -115,7 +125,7 @@ function generateFromTemplate({
   const currentVersionVarName = `${nextVersionCoerced.major}_${
     nextVersionCoerced.minor - 1
   }`;
-  return `import type { Profile, Package } from "../../types";
+  return `import type { Package, Profile } from "../../../types";
 import profile_${currentVersionVarName} from "./profile-${currentVersion}";
 
 const reactNative: Package = {
@@ -271,11 +281,11 @@ async function makeProfile(preset, targetVersion, latestProfile) {
  * @param {{ preset?: string; targetVersion?: string; force?: boolean; }} options
  */
 async function main({
-  preset: presetName = "microsoft",
+  preset: presetName = "microsoft/react-native",
   targetVersion = "",
   force,
 }) {
-  const { preset } = await import(`../lib/presets/${presetName}/index.js`);
+  const { preset } = await import(`../lib/presets/${presetName}.js`);
   const allVersions = /** @type {import("../src/types").ProfileVersion[]} */ (
     Object.keys(preset)
       .sort((lhs, rhs) => semverCompare(semverCoerce(lhs), semverCoerce(rhs)))
