+++ packages/align-deps/src/compatibility/config.ts
