+++ packages/align-deps/src/presets/microsoft/react-native/baseCapabilities.ts
