+++ packages/align-deps/src/presets/microsoft/react-native/profile-0.66.ts
