+++ packages/align-deps/test/check.app.test.ts
@@ -1,5 +1,5 @@
 import path from "path";
-import { checkPackageManifest } from "../src/check";
+import { checkPackageManifest, v2_checkPackageManifest } from "../src/check";
 
 jest.mock("fs");
 jest.unmock("@rnx-kit/config");
@@ -34,6 +34,31 @@ describe("checkPackageManifest({ kitType: 'app' })", () => {
       checkPackageManifest(manifestPath, { loose: false, write: true })
     ).toBe(0);
     expect(consoleWarnSpy).not.toBeCalled();
+    expect(destination).toBe(manifestPath);
+    expect(updatedManifest).toMatchSnapshot();
+  });
+});
+
+describe("v2_checkPackageManifest({ kitType: 'app' }) (backwards compatibility)", () => {
+  const fs = require("fs");
+
+  afterAll(() => {
+    jest.clearAllMocks();
+  });
+
+  test("adds required dependencies", () => {
+    const manifestPath = path.join(fixturePath("awesome-repo"), "package.json");
+
+    let destination = "";
+    let updatedManifest = "";
+    fs.__setMockFileWriter((dest, content) => {
+      destination = dest;
+      updatedManifest = content;
+    });
+
+    expect(
+      v2_checkPackageManifest(manifestPath, { loose: false, write: true })
+    ).toBe(0);
     expect(destination).toBe(manifestPath);
     expect(updatedManifest).toMatchSnapshot();
   });
