+++ packages/dep-check/src/presets/microsoft/baseCapabilities.ts
