+++ packages/dep-check/src/profiles.ts
@@ -6,7 +6,7 @@ import semverValid from "semver/functions/valid";
 import semverIntersects from "semver/ranges/intersects";
 import semverValidRange from "semver/ranges/valid";
 import { keysOf } from "./helpers";
-import { default as defaultPreset } from "./presets/microsoft/react-native";
+import { default as defaultPreset } from "./presets/microsoft";
 import type {
   MetaPackage,
   Package,
