+++ packages/dep-check/test/capabilities.test.ts
@@ -1,8 +1,8 @@
 import type { Capability } from "@rnx-kit/config";
 import { capabilitiesFor, resolveCapabilities } from "../src/capabilities";
-import profile_0_62 from "../src/presets/microsoft/react-native/profile-0.62";
-import profile_0_63 from "../src/presets/microsoft/react-native/profile-0.63";
-import profile_0_64 from "../src/presets/microsoft/react-native/profile-0.64";
+import profile_0_62 from "../src/presets/microsoft/profile-0.62";
+import profile_0_63 from "../src/presets/microsoft/profile-0.63";
+import profile_0_64 from "../src/presets/microsoft/profile-0.64";
 import { getProfilesFor } from "../src/profiles";
 import { pickPackage } from "./helpers";
