+++ packages/dep-check/test/profiles.test.ts
@@ -1,9 +1,9 @@
 import * as path from "path";
 import semver from "semver";
-import preset from "../src/presets/microsoft/react-native";
-import profile_0_62 from "../src/presets/microsoft/react-native/profile-0.62";
-import profile_0_63 from "../src/presets/microsoft/react-native/profile-0.63";
-import profile_0_64 from "../src/presets/microsoft/react-native/profile-0.64";
+import preset from "../src/presets/microsoft";
+import profile_0_62 from "../src/presets/microsoft/profile-0.62";
+import profile_0_63 from "../src/presets/microsoft/profile-0.63";
+import profile_0_64 from "../src/presets/microsoft/profile-0.64";
 import {
   getProfilesFor,
   getProfileVersionsFor,
@@ -11,11 +11,13 @@ import {
   profilesSatisfying,
   resolveCustomProfiles,
 } from "../src/profiles";
-import type { ProfileVersion } from "../src/types";
+import { ProfileVersion } from "../src/types";
 
-describe("microsoft/react-native preset", () => {
+describe("Microsoft preset", () => {
   test("matches react-native versions", () => {
-    const includePrerelease = { includePrerelease: true };
+    const includePrerelease = {
+      includePrerelease: true,
+    };
     Object.entries(preset).forEach(([version, capabilities]) => {
       const versionRange = "^" + version;
       Object.entries(capabilities).forEach(([capability, pkg]) => {
